fix: Image override content for small promo not working as expected
Use content hook should be top-level in functional components
